backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
bcrypt==4.2.1
pyjwt==2.10.1

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1

# Excel parsing
openpyxl==3.1.5
pandas==2.2.3
rapidfuzz==3.14.1

# Task queue
celery==5.4.0
redis==5.2.1

# Utils
python-multipart==0.0.19
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3

pdfplumber>=0.11.0
